foresight_old/requirements.txt
rejson==0.5.4
pytest>=6.1.1
fluent==0.10.0
requests==2.24.0
Pillow==8.0.1
ipython==7.18.1
namesgenerator==0.3
Rx==3.1.1 # Still used???
tuspy==0.2.4
Flask==1.1.2
python-magic==0.4.18
rq==1.5.2
dill==0.3.2
